requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic